clean up + restructuring of banner component
