+++ SwiftUIKit/BannerComponent2.0/BannerData.swift
