+++ SwiftUIKit/BannerComponent2.0/BannerManager.swift
