+++ SwiftUIKit/ContentView.swift
@@ -11,7 +11,7 @@ struct ContentView: View {
     
     var body: some View {
         if #available(iOS 16.0, *) {
-            BannerDemoView()
+            BannerSetUpView()
         } else {
             Text("No view")
         }
